# backend/docs/swagger.yaml
definitions:
  errs.Code:
    enum:
    - 100001
    - 100002
    - 100003
    - 100004
    type: integer
    x-enum-varnames:
    - ErrUnknown
    - ErrInvalidRequestParameters
    - ErrServerPanic
    - ErrServerExiting
...
  protocol.AppleAPIGetSigningJobInformationRsp:
    properties:
      bundleId:
        description: Bundle ID
        example: cn.ivfzhou.test
        type: string
      createdTime:
        description: 创建时间
        example: "2020-01-01 01:01:01"
        type: string
      fileId:
        description: 待签名文件 ID
        example: 2026014ef83c03e2ce4f1f94c11168d1acd087
        type: string
      fileName:
        description: 待签名文件名
        type: string
      finishedTime:
        description: 结束时间
        example: "2020-01-01 01:01:01"
        type: string
      log:
        description: 日志
        type: string
      profileId:
        description: 描述文件 ID
        example: 93aa1dc7f4ab32bdc5a3b0b78ecc35b5
        type: string
      signedFileId:
        description: 结果文件 ID
        example: 2026014ef83c03e2ce4f1f94c11168d1acd087
        type: string
      signedFileName:
        description: 结果文件名
        type: string
      source:
        description: 来源
        example: 1
        type: integer
      status:
        description: 状态
        example: 1
        type: integer
      userName:
        description: 提交人
        example: zhangsan
        type: string
    type: object
  protocol.AppleAPISubmitSigningJobReq:
    properties:
      fileId:
        description: 待文件 ID
        example: 2026014ef83c03e2ce4f1f94c11168d1acd087
        type: string
      profileId:
        description: 描述文件 ID
        example: 93aa1dc7f4ab32bdc5a3b0b78ecc35b5
        type: string
    type: object
  protocol.AppleAPISubmitSigningJobRsp:
    properties:
      jobId:
        description: 任务 ID
        example: 2026014ef83c03e2ce4f1f94c11168d1acd087
        type: string
    type: object
...
  util.Response-any:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
...
  util.Response-protocol_AppleAPIGetSigningJobInformationRsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        allOf:
        - $ref: '#/definitions/protocol.AppleAPIGetSigningJobInformationRsp'
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
  util.Response-protocol_AppleAPISubmitSigningJobRsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        allOf:
        - $ref: '#/definitions/protocol.AppleAPISubmitSigningJobRsp'
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
...
host: 127.0.0.1:8090
info:
  contact:
    email: ivfzhou@126.com
    name: ivfzhou
  license:
    name: MulanPSL-2.0 license
    url: http://license.coscl.org.cn/MulanPSL2
  title: CSMS API
  version: "1.0"
paths:
...
  /api/apple/downloadCertificate:
    get:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求凭据
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID
        example: 93aa1dc7f4ab32bdc5a3b0b78ecc35b5
        in: query
        name: certificateId
        type: string
      - description: 类型
        example: 1
        in: query
        maximum: 3
        minimum: 1
        name: type
        type: integer
      produces:
      - application/otect-stream
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              type: string
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response-any'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/util.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 下载证书和描述文件
      tags:
      - Apple-OpenAPI
  /api/apple/getSigningJobInformation:
    get:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求凭据
        in: header
        name: Authorization
        required: true
        type: string
      - description: 任务 ID
        example: 2026014ef83c03e2ce4f1f94c11168d1acd087
        in: query
        name: jobId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_AppleAPIGetSigningJobInformationRsp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response-any'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/util.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 获取签名任务信息
      tags:
      - Apple-OpenAPI
  /api/apple/submitSigningJob:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求凭据
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.AppleAPISubmitSigningJobReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_AppleAPISubmitSigningJobRsp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/util.Response-any'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/util.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 提交签名任务
      tags:
      - Apple-OpenAPI
...
schemes:
- http
swagger: "2.0"